feat lab7: 满分答案：前缀和要可能超过int范围，需要用64位整数

+++ src/main/cpp/cn/edu/SUSTech/YeCanming/Algs/lab7/problemA.cpp
@@ -7,7 +7,7 @@
 #include <vector>
 #include <limits>
 std::vector<uint32_t> weight;
-std::vector<uint32_t> _prefix_sum;
+std::vector<uint64_t> _prefix_sum;
 void prefix_sum(){
    _prefix_sum.push_back(0);
     for (size_t i = 0; i < weight.size(); i++)
